Changing the shift register clear routine
Was trying to track down a high current draw issue that was linked to
the shift registers ending up in some weird state. Modified the clear()
function. Also modified read16Hall to perform additional shift register
clears and to explicitly pull the multiplexer ENABLE pin high to disable
it after a cycle of sampling.